backend/pyproject.toml
[project]
name = "videonote-backend"
version = "0.1.0"
description = "VideoNote 后端 - 视频转笔记 API 服务"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.50.0",
    "faster-whisper>=1.1.0",
    "yt-dlp>=2024.12.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
